+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - skipping model loading for disabled providers
 - updated visuals of model selectors
 - Claude Agent SDK provider only available when claude executable is present
+- added action to handoff a todo item to a new task
 
 ## [0.50.1]